requirements.txt
asgiref==3.9.1
attrs==25.3.0
django==5.2.5
djangorestframework==3.16.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
pip==25.2
psycopg2==2.9.10
pyyaml==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0